ISSUP-3469 remove never used trial type, SF

+++ modules/dxp/apps/osb/osb-provisioning/osb-provisioning-license/osb-provisioning-license-service/src/main/java/com/liferay/osb/provisioning/license/service/impl/LicenseKeyLocalServiceImpl.java
@@ -465,8 +465,8 @@ public class LicenseKeyLocalServiceImpl extends LicenseKeyLocalServiceBaseImpl {
 
 		licenseKey = licenseKeyPersistence.update(licenseKey);
 
-		int renewTime =
-			(int)((expirationDate.getTime() - startDate.getTime()) / Time.DAY);
+		long renewTime =
+			(expirationDate.getTime() - startDate.getTime()) / Time.DAY;
 
 		String description = renewTime + "-Day License";
